Crate swc_ecma_visit

Modules

fields

Macros

assert_eq_ignore_span

noop_fold_type

Note: Ignoring more types is not considered as a breaking change.

noop_visit_mut_type

Note: Ignoring more types is not considered as a breaking change.

noop_visit_type

Note: Ignoring more types is not considered as a breaking change.

standard_only_fold

Mark visitor as ECMAScript standard only and mark other types as unreachable.

standard_only_visit

Mark visitor as ECMAScript standard only and mark other types as unreachable.

standard_only_visit_mut

Mark visitor as ECMAScript standard only and mark other types as unreachable.

visit_mut_obj_and_computed

visit_obj_and_computed

Structs

All

Visit all children nodes. This converts VisitAll to Visit. The type parameter V should implement VisitAll and All<V> implements Visit.

Folder

Wrap a VisitMut as a Fold

Enums

AstParentKind

AstParentNodeRef

NodeRef

Traits

Fold

A visitor trait for traversing the AST.

FoldAstPathpath

A visitor trait for traversing the AST.

FoldWith

A trait implemented for types that can be visited using a visitor.

FoldWithAstPathpath

A trait implemented for types that can be visited using a visitor.

InjectVars

Implemented for passes which inject variables.

Visit

A visitor trait for traversing the AST.

VisitAstPathpath

A visitor trait for traversing the AST.

VisitMut

A visitor trait for traversing the AST.

VisitMutAstPathpath

A visitor trait for traversing the AST.

VisitMutWith

A trait implemented for types that can be visited using a visitor.

VisitMutWithAstPathpath

A trait implemented for types that can be visited using a visitor.

VisitWith

A trait implemented for types that can be visited using a visitor.

VisitWithAstPathpath

A trait implemented for types that can be visited using a visitor.

Functions

as_folder

The returned folder only handles fold_script and fold_module, and typescript nodes are ignored. So if your visitor needs to handle typescript or low-level nodes, you should use as_folder instead.

span_remover

Returns a Fold which changes all span into DUMMY_SP.

Type Aliases

AstKindPath

AstNodePath