swc_ecma_visit

Trait InjectVars

pub trait InjectVars {
    // Required method
    fn take_vars(&mut self) -> Vec<VarDeclarator>;
}

Implemented for passes which inject variables.

If a pass depends on other pass which injects variables, this trait can be used to keep the variables.

Required Methods

fn take_vars(&mut self) -> Vec<VarDeclarator>