Trait swc_ecma_visit::VisitAll

pub trait VisitAll {

    fn visit_accessibility(&mut self, n: &Accessibility) { ... }
    fn visit_array_lit(&mut self, n: &ArrayLit) { ... }
    fn visit_array_pat(&mut self, n: &ArrayPat) { ... }
    fn visit_arrow_expr(&mut self, n: &ArrowExpr) { ... }
    fn visit_assign_expr(&mut self, n: &AssignExpr) { ... }
    fn visit_assign_op(&mut self, n: &AssignOp) { ... }
    fn visit_assign_pat(&mut self, n: &AssignPat) { ... }
    fn visit_assign_pat_prop(&mut self, n: &AssignPatProp) { ... }
    fn visit_assign_prop(&mut self, n: &AssignProp) { ... }
    fn visit_atom(&mut self, n: &Atom) { ... }
    fn visit_await_expr(&mut self, n: &AwaitExpr) { ... }
    fn visit_big_int(&mut self, n: &BigInt) { ... }
    fn visit_big_int_value(&mut self, n: &BigIntValue) { ... }
    fn visit_bin_expr(&mut self, n: &BinExpr) { ... }
    fn visit_binary_op(&mut self, n: &BinaryOp) { ... }
    fn visit_binding_ident(&mut self, n: &BindingIdent) { ... }
    fn visit_block_stmt(&mut self, n: &BlockStmt) { ... }
    fn visit_block_stmt_or_expr(&mut self, n: &BlockStmtOrExpr) { ... }
    fn visit_bool(&mut self, n: &Bool) { ... }
    fn visit_break_stmt(&mut self, n: &BreakStmt) { ... }
    fn visit_call_expr(&mut self, n: &CallExpr) { ... }
    fn visit_callee(&mut self, n: &Callee) { ... }
    fn visit_catch_clause(&mut self, n: &CatchClause) { ... }
    fn visit_class(&mut self, n: &Class) { ... }
    fn visit_class_decl(&mut self, n: &ClassDecl) { ... }
    fn visit_class_expr(&mut self, n: &ClassExpr) { ... }
    fn visit_class_member(&mut self, n: &ClassMember) { ... }
    fn visit_class_members(&mut self, n: &[ClassMember]) { ... }
    fn visit_class_method(&mut self, n: &ClassMethod) { ... }
    fn visit_class_prop(&mut self, n: &ClassProp) { ... }
    fn visit_computed_prop_name(&mut self, n: &ComputedPropName) { ... }
    fn visit_cond_expr(&mut self, n: &CondExpr) { ... }
    fn visit_constructor(&mut self, n: &Constructor) { ... }
    fn visit_continue_stmt(&mut self, n: &ContinueStmt) { ... }
    fn visit_debugger_stmt(&mut self, n: &DebuggerStmt) { ... }
    fn visit_decl(&mut self, n: &Decl) { ... }
    fn visit_decorator(&mut self, n: &Decorator) { ... }
    fn visit_decorators(&mut self, n: &[Decorator]) { ... }
    fn visit_default_decl(&mut self, n: &DefaultDecl) { ... }
    fn visit_do_while_stmt(&mut self, n: &DoWhileStmt) { ... }
    fn visit_empty_stmt(&mut self, n: &EmptyStmt) { ... }
    fn visit_export_all(&mut self, n: &ExportAll) { ... }
    fn visit_export_decl(&mut self, n: &ExportDecl) { ... }
    fn visit_export_default_decl(&mut self, n: &ExportDefaultDecl) { ... }
    fn visit_export_default_expr(&mut self, n: &ExportDefaultExpr) { ... }
    fn visit_export_default_specifier(&mut self, n: &ExportDefaultSpecifier) { ... }
    fn visit_export_named_specifier(&mut self, n: &ExportNamedSpecifier) { ... }
    fn visit_export_namespace_specifier(&mut self, n: &ExportNamespaceSpecifier) { ... }
    fn visit_export_specifier(&mut self, n: &ExportSpecifier) { ... }
    fn visit_export_specifiers(&mut self, n: &[ExportSpecifier]) { ... }
    fn visit_expr(&mut self, n: &Expr) { ... }
    fn visit_expr_or_spread(&mut self, n: &ExprOrSpread) { ... }
    fn visit_expr_or_spreads(&mut self, n: &[ExprOrSpread]) { ... }
    fn visit_expr_stmt(&mut self, n: &ExprStmt) { ... }
    fn visit_exprs(&mut self, n: &[Box<Expr>]) { ... }
    fn visit_fn_decl(&mut self, n: &FnDecl) { ... }
    fn visit_fn_expr(&mut self, n: &FnExpr) { ... }
    fn visit_for_in_stmt(&mut self, n: &ForInStmt) { ... }
    fn visit_for_of_stmt(&mut self, n: &ForOfStmt) { ... }
    fn visit_for_stmt(&mut self, n: &ForStmt) { ... }
    fn visit_function(&mut self, n: &Function) { ... }
    fn visit_getter_prop(&mut self, n: &GetterProp) { ... }
    fn visit_ident(&mut self, n: &Ident) { ... }
    fn visit_if_stmt(&mut self, n: &IfStmt) { ... }
    fn visit_import(&mut self, n: &Import) { ... }
    fn visit_import_decl(&mut self, n: &ImportDecl) { ... }
    fn visit_import_default_specifier(&mut self, n: &ImportDefaultSpecifier) { ... }
    fn visit_import_named_specifier(&mut self, n: &ImportNamedSpecifier) { ... }
    fn visit_import_specifier(&mut self, n: &ImportSpecifier) { ... }
    fn visit_import_specifiers(&mut self, n: &[ImportSpecifier]) { ... }
    fn visit_import_star_as_specifier(&mut self, n: &ImportStarAsSpecifier) { ... }
    fn visit_invalid(&mut self, n: &Invalid) { ... }
    fn visit_js_word(&mut self, n: &JsWord) { ... }
    fn visit_jsx_attr(&mut self, n: &JSXAttr) { ... }
    fn visit_jsx_attr_name(&mut self, n: &JSXAttrName) { ... }
    fn visit_jsx_attr_or_spread(&mut self, n: &JSXAttrOrSpread) { ... }
    fn visit_jsx_attr_or_spreads(&mut self, n: &[JSXAttrOrSpread]) { ... }
    fn visit_jsx_attr_value(&mut self, n: &JSXAttrValue) { ... }
    fn visit_jsx_closing_element(&mut self, n: &JSXClosingElement) { ... }
    fn visit_jsx_closing_fragment(&mut self, n: &JSXClosingFragment) { ... }
    fn visit_jsx_element(&mut self, n: &JSXElement) { ... }
    fn visit_jsx_element_child(&mut self, n: &JSXElementChild) { ... }
    fn visit_jsx_element_children(&mut self, n: &[JSXElementChild]) { ... }
    fn visit_jsx_element_name(&mut self, n: &JSXElementName) { ... }
    fn visit_jsx_empty_expr(&mut self, n: &JSXEmptyExpr) { ... }
    fn visit_jsx_expr(&mut self, n: &JSXExpr) { ... }
    fn visit_jsx_expr_container(&mut self, n: &JSXExprContainer) { ... }
    fn visit_jsx_fragment(&mut self, n: &JSXFragment) { ... }
    fn visit_jsx_member_expr(&mut self, n: &JSXMemberExpr) { ... }
    fn visit_jsx_namespaced_name(&mut self, n: &JSXNamespacedName) { ... }
    fn visit_jsx_object(&mut self, n: &JSXObject) { ... }
    fn visit_jsx_opening_element(&mut self, n: &JSXOpeningElement) { ... }
    fn visit_jsx_opening_fragment(&mut self, n: &JSXOpeningFragment) { ... }
    fn visit_jsx_spread_child(&mut self, n: &JSXSpreadChild) { ... }
    fn visit_jsx_text(&mut self, n: &JSXText) { ... }
    fn visit_key_value_pat_prop(&mut self, n: &KeyValuePatProp) { ... }
    fn visit_key_value_prop(&mut self, n: &KeyValueProp) { ... }
    fn visit_labeled_stmt(&mut self, n: &LabeledStmt) { ... }
    fn visit_lit(&mut self, n: &Lit) { ... }
    fn visit_member_expr(&mut self, n: &MemberExpr) { ... }
    fn visit_member_prop(&mut self, n: &MemberProp) { ... }
    fn visit_meta_prop_expr(&mut self, n: &MetaPropExpr) { ... }
    fn visit_meta_prop_kind(&mut self, n: &MetaPropKind) { ... }
    fn visit_method_kind(&mut self, n: &MethodKind) { ... }
    fn visit_method_prop(&mut self, n: &MethodProp) { ... }
    fn visit_module(&mut self, n: &Module) { ... }
    fn visit_module_decl(&mut self, n: &ModuleDecl) { ... }
    fn visit_module_export_name(&mut self, n: &ModuleExportName) { ... }
    fn visit_module_item(&mut self, n: &ModuleItem) { ... }
    fn visit_module_items(&mut self, n: &[ModuleItem]) { ... }
    fn visit_named_export(&mut self, n: &NamedExport) { ... }
    fn visit_new_expr(&mut self, n: &NewExpr) { ... }
    fn visit_null(&mut self, n: &Null) { ... }
    fn visit_number(&mut self, n: &Number) { ... }
    fn visit_object_lit(&mut self, n: &ObjectLit) { ... }
    fn visit_object_pat(&mut self, n: &ObjectPat) { ... }
    fn visit_object_pat_prop(&mut self, n: &ObjectPatProp) { ... }
    fn visit_object_pat_props(&mut self, n: &[ObjectPatProp]) { ... }
    fn visit_opt_accessibility(&mut self, n: Option<&Accessibility>) { ... }
    fn visit_opt_atom(&mut self, n: Option<&Atom>) { ... }
    fn visit_opt_block_stmt(&mut self, n: Option<&BlockStmt>) { ... }
    fn visit_opt_call(&mut self, n: &OptCall) { ... }
    fn visit_opt_catch_clause(&mut self, n: Option<&CatchClause>) { ... }
    fn visit_opt_chain_base(&mut self, n: &OptChainBase) { ... }
    fn visit_opt_chain_expr(&mut self, n: &OptChainExpr) { ... }
    fn visit_opt_expr(&mut self, n: Option<&Box<Expr>>) { ... }
    fn visit_opt_expr_or_spread(&mut self, n: Option<&ExprOrSpread>) { ... }
    fn visit_opt_expr_or_spreads(&mut self, n: Option<&[ExprOrSpread]>) { ... }
    fn visit_opt_ident(&mut self, n: Option<&Ident>) { ... }
    fn visit_opt_jsx_attr_value(&mut self, n: Option<&JSXAttrValue>) { ... }
    fn visit_opt_jsx_closing_element(&mut self, n: Option<&JSXClosingElement>) { ... }
    fn visit_opt_module_export_name(&mut self, n: Option<&ModuleExportName>) { ... }
    fn visit_opt_module_items(&mut self, n: Option<&[ModuleItem]>) { ... }
    fn visit_opt_object_lit(&mut self, n: Option<&Box<ObjectLit>>) { ... }
    fn visit_opt_pat(&mut self, n: Option<&Pat>) { ... }
    fn visit_opt_span(&mut self, n: Option<&Span>) { ... }
    fn visit_opt_stmt(&mut self, n: Option<&Box<Stmt>>) { ... }
    fn visit_opt_str(&mut self, n: Option<&Box<Str>>) { ... }
    fn visit_opt_true_plus_minus(&mut self, n: Option<&TruePlusMinus>) { ... }
    fn visit_opt_ts_entity_name(&mut self, n: Option<&TsEntityName>) { ... }
    fn visit_opt_ts_namespace_body(&mut self, n: Option<&TsNamespaceBody>) { ... }
    fn visit_opt_ts_type(&mut self, n: Option<&Box<TsType>>) { ... }
    fn visit_opt_ts_type_ann(&mut self, n: Option<&Box<TsTypeAnn>>) { ... }
    fn visit_opt_ts_type_param_decl(&mut self, n: Option<&Box<TsTypeParamDecl>>) { ... }
    fn visit_opt_ts_type_param_instantiation(
        &mut self,
        n: Option<&Box<TsTypeParamInstantiation>>
    ) { ... }
    fn visit_opt_var_decl_or_expr(&mut self, n: Option<&VarDeclOrExpr>) { ... }
    fn visit_opt_vec_expr_or_spreads(&mut self, n: &[Option<ExprOrSpread>]) { ... }
    fn visit_opt_vec_pats(&mut self, n: &[Option<Pat>]) { ... }
    fn visit_param(&mut self, n: &Param) { ... }
    fn visit_param_or_ts_param_prop(&mut self, n: &ParamOrTsParamProp) { ... }
    fn visit_param_or_ts_param_props(&mut self, n: &[ParamOrTsParamProp]) { ... }
    fn visit_params(&mut self, n: &[Param]) { ... }
    fn visit_paren_expr(&mut self, n: &ParenExpr) { ... }
    fn visit_pat(&mut self, n: &Pat) { ... }
    fn visit_pat_or_expr(&mut self, n: &PatOrExpr) { ... }
    fn visit_pats(&mut self, n: &[Pat]) { ... }
    fn visit_private_method(&mut self, n: &PrivateMethod) { ... }
    fn visit_private_name(&mut self, n: &PrivateName) { ... }
    fn visit_private_prop(&mut self, n: &PrivateProp) { ... }
    fn visit_program(&mut self, n: &Program) { ... }
    fn visit_prop(&mut self, n: &Prop) { ... }
    fn visit_prop_name(&mut self, n: &PropName) { ... }
    fn visit_prop_or_spread(&mut self, n: &PropOrSpread) { ... }
    fn visit_prop_or_spreads(&mut self, n: &[PropOrSpread]) { ... }
    fn visit_regex(&mut self, n: &Regex) { ... }
    fn visit_reserved_unused(&mut self, n: &ReservedUnused) { ... }
    fn visit_rest_pat(&mut self, n: &RestPat) { ... }
    fn visit_return_stmt(&mut self, n: &ReturnStmt) { ... }
    fn visit_script(&mut self, n: &Script) { ... }
    fn visit_seq_expr(&mut self, n: &SeqExpr) { ... }
    fn visit_setter_prop(&mut self, n: &SetterProp) { ... }
    fn visit_span(&mut self, n: &Span) { ... }
    fn visit_spread_element(&mut self, n: &SpreadElement) { ... }
    fn visit_static_block(&mut self, n: &StaticBlock) { ... }
    fn visit_stmt(&mut self, n: &Stmt) { ... }
    fn visit_stmts(&mut self, n: &[Stmt]) { ... }
    fn visit_str(&mut self, n: &Str) { ... }
    fn visit_super(&mut self, n: &Super) { ... }
    fn visit_super_prop(&mut self, n: &SuperProp) { ... }
    fn visit_super_prop_expr(&mut self, n: &SuperPropExpr) { ... }
    fn visit_switch_case(&mut self, n: &SwitchCase) { ... }
    fn visit_switch_cases(&mut self, n: &[SwitchCase]) { ... }
    fn visit_switch_stmt(&mut self, n: &SwitchStmt) { ... }
    fn visit_tagged_tpl(&mut self, n: &TaggedTpl) { ... }
    fn visit_this_expr(&mut self, n: &ThisExpr) { ... }
    fn visit_throw_stmt(&mut self, n: &ThrowStmt) { ... }
    fn visit_tpl(&mut self, n: &Tpl) { ... }
    fn visit_tpl_element(&mut self, n: &TplElement) { ... }
    fn visit_tpl_elements(&mut self, n: &[TplElement]) { ... }
    fn visit_true_plus_minus(&mut self, n: &TruePlusMinus) { ... }
    fn visit_try_stmt(&mut self, n: &TryStmt) { ... }
    fn visit_ts_array_type(&mut self, n: &TsArrayType) { ... }
    fn visit_ts_as_expr(&mut self, n: &TsAsExpr) { ... }
    fn visit_ts_call_signature_decl(&mut self, n: &TsCallSignatureDecl) { ... }
    fn visit_ts_conditional_type(&mut self, n: &TsConditionalType) { ... }
    fn visit_ts_const_assertion(&mut self, n: &TsConstAssertion) { ... }
    fn visit_ts_construct_signature_decl(&mut self, n: &TsConstructSignatureDecl) { ... }
    fn visit_ts_constructor_type(&mut self, n: &TsConstructorType) { ... }
    fn visit_ts_entity_name(&mut self, n: &TsEntityName) { ... }
    fn visit_ts_enum_decl(&mut self, n: &TsEnumDecl) { ... }
    fn visit_ts_enum_member(&mut self, n: &TsEnumMember) { ... }
    fn visit_ts_enum_member_id(&mut self, n: &TsEnumMemberId) { ... }
    fn visit_ts_enum_members(&mut self, n: &[TsEnumMember]) { ... }
    fn visit_ts_export_assignment(&mut self, n: &TsExportAssignment) { ... }
    fn visit_ts_expr_with_type_args(&mut self, n: &TsExprWithTypeArgs) { ... }
    fn visit_ts_expr_with_type_args_vec(&mut self, n: &[TsExprWithTypeArgs]) { ... }
    fn visit_ts_external_module_ref(&mut self, n: &TsExternalModuleRef) { ... }
    fn visit_ts_fn_or_constructor_type(&mut self, n: &TsFnOrConstructorType) { ... }
    fn visit_ts_fn_param(&mut self, n: &TsFnParam) { ... }
    fn visit_ts_fn_params(&mut self, n: &[TsFnParam]) { ... }
    fn visit_ts_fn_type(&mut self, n: &TsFnType) { ... }
    fn visit_ts_getter_signature(&mut self, n: &TsGetterSignature) { ... }
    fn visit_ts_import_equals_decl(&mut self, n: &TsImportEqualsDecl) { ... }
    fn visit_ts_import_type(&mut self, n: &TsImportType) { ... }
    fn visit_ts_index_signature(&mut self, n: &TsIndexSignature) { ... }
    fn visit_ts_indexed_access_type(&mut self, n: &TsIndexedAccessType) { ... }
    fn visit_ts_infer_type(&mut self, n: &TsInferType) { ... }
    fn visit_ts_instantiation(&mut self, n: &TsInstantiation) { ... }
    fn visit_ts_interface_body(&mut self, n: &TsInterfaceBody) { ... }
    fn visit_ts_interface_decl(&mut self, n: &TsInterfaceDecl) { ... }
    fn visit_ts_intersection_type(&mut self, n: &TsIntersectionType) { ... }
    fn visit_ts_keyword_type(&mut self, n: &TsKeywordType) { ... }
    fn visit_ts_keyword_type_kind(&mut self, n: &TsKeywordTypeKind) { ... }
    fn visit_ts_lit(&mut self, n: &TsLit) { ... }
    fn visit_ts_lit_type(&mut self, n: &TsLitType) { ... }
    fn visit_ts_mapped_type(&mut self, n: &TsMappedType) { ... }
    fn visit_ts_method_signature(&mut self, n: &TsMethodSignature) { ... }
    fn visit_ts_module_block(&mut self, n: &TsModuleBlock) { ... }
    fn visit_ts_module_decl(&mut self, n: &TsModuleDecl) { ... }
    fn visit_ts_module_name(&mut self, n: &TsModuleName) { ... }
    fn visit_ts_module_ref(&mut self, n: &TsModuleRef) { ... }
    fn visit_ts_namespace_body(&mut self, n: &TsNamespaceBody) { ... }
    fn visit_ts_namespace_decl(&mut self, n: &TsNamespaceDecl) { ... }
    fn visit_ts_namespace_export_decl(&mut self, n: &TsNamespaceExportDecl) { ... }
    fn visit_ts_non_null_expr(&mut self, n: &TsNonNullExpr) { ... }
    fn visit_ts_optional_type(&mut self, n: &TsOptionalType) { ... }
    fn visit_ts_param_prop(&mut self, n: &TsParamProp) { ... }
    fn visit_ts_param_prop_param(&mut self, n: &TsParamPropParam) { ... }
    fn visit_ts_parenthesized_type(&mut self, n: &TsParenthesizedType) { ... }
    fn visit_ts_property_signature(&mut self, n: &TsPropertySignature) { ... }
    fn visit_ts_qualified_name(&mut self, n: &TsQualifiedName) { ... }
    fn visit_ts_rest_type(&mut self, n: &TsRestType) { ... }
    fn visit_ts_satisfies_expr(&mut self, n: &TsSatisfiesExpr) { ... }
    fn visit_ts_setter_signature(&mut self, n: &TsSetterSignature) { ... }
    fn visit_ts_this_type(&mut self, n: &TsThisType) { ... }
    fn visit_ts_this_type_or_ident(&mut self, n: &TsThisTypeOrIdent) { ... }
    fn visit_ts_tpl_lit_type(&mut self, n: &TsTplLitType) { ... }
    fn visit_ts_tuple_element(&mut self, n: &TsTupleElement) { ... }
    fn visit_ts_tuple_elements(&mut self, n: &[TsTupleElement]) { ... }
    fn visit_ts_tuple_type(&mut self, n: &TsTupleType) { ... }
    fn visit_ts_type(&mut self, n: &TsType) { ... }
    fn visit_ts_type_alias_decl(&mut self, n: &TsTypeAliasDecl) { ... }
    fn visit_ts_type_ann(&mut self, n: &TsTypeAnn) { ... }
    fn visit_ts_type_assertion(&mut self, n: &TsTypeAssertion) { ... }
    fn visit_ts_type_element(&mut self, n: &TsTypeElement) { ... }
    fn visit_ts_type_elements(&mut self, n: &[TsTypeElement]) { ... }
    fn visit_ts_type_lit(&mut self, n: &TsTypeLit) { ... }
    fn visit_ts_type_operator(&mut self, n: &TsTypeOperator) { ... }
    fn visit_ts_type_operator_op(&mut self, n: &TsTypeOperatorOp) { ... }
    fn visit_ts_type_param(&mut self, n: &TsTypeParam) { ... }
    fn visit_ts_type_param_decl(&mut self, n: &TsTypeParamDecl) { ... }
    fn visit_ts_type_param_instantiation(&mut self, n: &TsTypeParamInstantiation) { ... }
    fn visit_ts_type_params(&mut self, n: &[TsTypeParam]) { ... }
    fn visit_ts_type_predicate(&mut self, n: &TsTypePredicate) { ... }
    fn visit_ts_type_query(&mut self, n: &TsTypeQuery) { ... }
    fn visit_ts_type_query_expr(&mut self, n: &TsTypeQueryExpr) { ... }
    fn visit_ts_type_ref(&mut self, n: &TsTypeRef) { ... }
    fn visit_ts_types(&mut self, n: &[Box<TsType>]) { ... }
    fn visit_ts_union_or_intersection_type(
        &mut self,
        n: &TsUnionOrIntersectionType
    ) { ... }
    fn visit_ts_union_type(&mut self, n: &TsUnionType) { ... }
    fn visit_unary_expr(&mut self, n: &UnaryExpr) { ... }
    fn visit_unary_op(&mut self, n: &UnaryOp) { ... }
    fn visit_update_expr(&mut self, n: &UpdateExpr) { ... }
    fn visit_update_op(&mut self, n: &UpdateOp) { ... }
    fn visit_var_decl(&mut self, n: &VarDecl) { ... }
    fn visit_var_decl_kind(&mut self, n: &VarDeclKind) { ... }
    fn visit_var_decl_or_expr(&mut self, n: &VarDeclOrExpr) { ... }
    fn visit_var_decl_or_pat(&mut self, n: &VarDeclOrPat) { ... }
    fn visit_var_declarator(&mut self, n: &VarDeclarator) { ... }
    fn visit_var_declarators(&mut self, n: &[VarDeclarator]) { ... }
    fn visit_while_stmt(&mut self, n: &WhileStmt) { ... }
    fn visit_with_stmt(&mut self, n: &WithStmt) { ... }
    fn visit_yield_expr(&mut self, n: &YieldExpr) { ... }

}

Provided Methods

fn visit_accessibility(&mut self, n: &Accessibility)

This method can be overriden to customize the visitor behavior.

fn visit_array_lit(&mut self, n: &ArrayLit)

This method can be overriden to customize the visitor behavior.

fn visit_array_pat(&mut self, n: &ArrayPat)

This method can be overriden to customize the visitor behavior.

fn visit_arrow_expr(&mut self, n: &ArrowExpr)

This method can be overriden to customize the visitor behavior.

fn visit_assign_expr(&mut self, n: &AssignExpr)

This method can be overriden to customize the visitor behavior.

fn visit_assign_op(&mut self, n: &AssignOp)

This method can be overriden to customize the visitor behavior.

fn visit_assign_pat(&mut self, n: &AssignPat)

This method can be overriden to customize the visitor behavior.

fn visit_assign_pat_prop(&mut self, n: &AssignPatProp)

This method can be overriden to customize the visitor behavior.

fn visit_assign_prop(&mut self, n: &AssignProp)

This method can be overriden to customize the visitor behavior.

fn visit_atom(&mut self, n: &Atom)

fn visit_await_expr(&mut self, n: &AwaitExpr)

This method can be overriden to customize the visitor behavior.

fn visit_big_int(&mut self, n: &BigInt)

This method can be overriden to customize the visitor behavior.

fn visit_big_int_value(&mut self, n: &BigIntValue)

fn visit_bin_expr(&mut self, n: &BinExpr)

This method can be overriden to customize the visitor behavior.

fn visit_binary_op(&mut self, n: &BinaryOp)

This method can be overriden to customize the visitor behavior.

fn visit_binding_ident(&mut self, n: &BindingIdent)

This method can be overriden to customize the visitor behavior.

fn visit_block_stmt(&mut self, n: &BlockStmt)

This method can be overriden to customize the visitor behavior.

fn visit_block_stmt_or_expr(&mut self, n: &BlockStmtOrExpr)

This method can be overriden to customize the visitor behavior.

fn visit_bool(&mut self, n: &Bool)

This method can be overriden to customize the visitor behavior.

fn visit_break_stmt(&mut self, n: &BreakStmt)

This method can be overriden to customize the visitor behavior.

fn visit_call_expr(&mut self, n: &CallExpr)

This method can be overriden to customize the visitor behavior.

fn visit_callee(&mut self, n: &Callee)

This method can be overriden to customize the visitor behavior.

fn visit_catch_clause(&mut self, n: &CatchClause)

This method can be overriden to customize the visitor behavior.

fn visit_class(&mut self, n: &Class)

This method can be overriden to customize the visitor behavior.

fn visit_class_decl(&mut self, n: &ClassDecl)

This method can be overriden to customize the visitor behavior.

fn visit_class_expr(&mut self, n: &ClassExpr)

This method can be overriden to customize the visitor behavior.

fn visit_class_member(&mut self, n: &ClassMember)

This method can be overriden to customize the visitor behavior.

fn visit_class_members(&mut self, n: &[ClassMember])

fn visit_class_method(&mut self, n: &ClassMethod)

This method can be overriden to customize the visitor behavior.

fn visit_class_prop(&mut self, n: &ClassProp)

This method can be overriden to customize the visitor behavior.

fn visit_computed_prop_name(&mut self, n: &ComputedPropName)

This method can be overriden to customize the visitor behavior.

fn visit_cond_expr(&mut self, n: &CondExpr)

This method can be overriden to customize the visitor behavior.

fn visit_constructor(&mut self, n: &Constructor)

This method can be overriden to customize the visitor behavior.

fn visit_continue_stmt(&mut self, n: &ContinueStmt)

This method can be overriden to customize the visitor behavior.

fn visit_debugger_stmt(&mut self, n: &DebuggerStmt)

This method can be overriden to customize the visitor behavior.

fn visit_decl(&mut self, n: &Decl)

This method can be overriden to customize the visitor behavior.

fn visit_decorator(&mut self, n: &Decorator)

This method can be overriden to customize the visitor behavior.

fn visit_decorators(&mut self, n: &[Decorator])

fn visit_default_decl(&mut self, n: &DefaultDecl)

This method can be overriden to customize the visitor behavior.

fn visit_do_while_stmt(&mut self, n: &DoWhileStmt)

This method can be overriden to customize the visitor behavior.

fn visit_empty_stmt(&mut self, n: &EmptyStmt)

This method can be overriden to customize the visitor behavior.

fn visit_export_all(&mut self, n: &ExportAll)

This method can be overriden to customize the visitor behavior.

fn visit_export_decl(&mut self, n: &ExportDecl)

This method can be overriden to customize the visitor behavior.

fn visit_export_default_decl(&mut self, n: &ExportDefaultDecl)

This method can be overriden to customize the visitor behavior.

fn visit_export_default_expr(&mut self, n: &ExportDefaultExpr)

This method can be overriden to customize the visitor behavior.

fn visit_export_default_specifier(&mut self, n: &ExportDefaultSpecifier)

This method can be overriden to customize the visitor behavior.

fn visit_export_named_specifier(&mut self, n: &ExportNamedSpecifier)

This method can be overriden to customize the visitor behavior.

fn visit_export_namespace_specifier(&mut self, n: &ExportNamespaceSpecifier)

This method can be overriden to customize the visitor behavior.

fn visit_export_specifier(&mut self, n: &ExportSpecifier)

This method can be overriden to customize the visitor behavior.

fn visit_export_specifiers(&mut self, n: &[ExportSpecifier])

fn visit_expr(&mut self, n: &Expr)

This method can be overriden to customize the visitor behavior.

fn visit_expr_or_spread(&mut self, n: &ExprOrSpread)

This method can be overriden to customize the visitor behavior.

fn visit_expr_or_spreads(&mut self, n: &[ExprOrSpread])

fn visit_expr_stmt(&mut self, n: &ExprStmt)

This method can be overriden to customize the visitor behavior.

fn visit_exprs(&mut self, n: &[Box<Expr>])

fn visit_fn_decl(&mut self, n: &FnDecl)

This method can be overriden to customize the visitor behavior.

fn visit_fn_expr(&mut self, n: &FnExpr)

This method can be overriden to customize the visitor behavior.

fn visit_for_in_stmt(&mut self, n: &ForInStmt)

This method can be overriden to customize the visitor behavior.

fn visit_for_of_stmt(&mut self, n: &ForOfStmt)

This method can be overriden to customize the visitor behavior.

fn visit_for_stmt(&mut self, n: &ForStmt)

This method can be overriden to customize the visitor behavior.

fn visit_function(&mut self, n: &Function)

This method can be overriden to customize the visitor behavior.

fn visit_getter_prop(&mut self, n: &GetterProp)

This method can be overriden to customize the visitor behavior.

fn visit_ident(&mut self, n: &Ident)

This method can be overriden to customize the visitor behavior.

fn visit_if_stmt(&mut self, n: &IfStmt)

This method can be overriden to customize the visitor behavior.

fn visit_import(&mut self, n: &Import)

This method can be overriden to customize the visitor behavior.

fn visit_import_decl(&mut self, n: &ImportDecl)

This method can be overriden to customize the visitor behavior.

fn visit_import_default_specifier(&mut self, n: &ImportDefaultSpecifier)

This method can be overriden to customize the visitor behavior.

fn visit_import_named_specifier(&mut self, n: &ImportNamedSpecifier)

This method can be overriden to customize the visitor behavior.

fn visit_import_specifier(&mut self, n: &ImportSpecifier)

This method can be overriden to customize the visitor behavior.

fn visit_import_specifiers(&mut self, n: &[ImportSpecifier])

fn visit_import_star_as_specifier(&mut self, n: &ImportStarAsSpecifier)

This method can be overriden to customize the visitor behavior.

fn visit_invalid(&mut self, n: &Invalid)

This method can be overriden to customize the visitor behavior.

fn visit_js_word(&mut self, n: &JsWord)

fn visit_jsx_attr(&mut self, n: &JSXAttr)

This method can be overriden to customize the visitor behavior.

fn visit_jsx_attr_name(&mut self, n: &JSXAttrName)

This method can be overriden to customize the visitor behavior.

fn visit_jsx_attr_or_spread(&mut self, n: &JSXAttrOrSpread)

This method can be overriden to customize the visitor behavior.

fn visit_jsx_attr_or_spreads(&mut self, n: &[JSXAttrOrSpread])

fn visit_jsx_attr_value(&mut self, n: &JSXAttrValue)

This method can be overriden to customize the visitor behavior.

fn visit_jsx_closing_element(&mut self, n: &JSXClosingElement)

This method can be overriden to customize the visitor behavior.

fn visit_jsx_closing_fragment(&mut self, n: &JSXClosingFragment)

This method can be overriden to customize the visitor behavior.

fn visit_jsx_element(&mut self, n: &JSXElement)

This method can be overriden to customize the visitor behavior.

fn visit_jsx_element_child(&mut self, n: &JSXElementChild)

This method can be overriden to customize the visitor behavior.

fn visit_jsx_element_children(&mut self, n: &[JSXElementChild])

fn visit_jsx_element_name(&mut self, n: &JSXElementName)

This method can be overriden to customize the visitor behavior.

fn visit_jsx_empty_expr(&mut self, n: &JSXEmptyExpr)

This method can be overriden to customize the visitor behavior.

fn visit_jsx_expr(&mut self, n: &JSXExpr)

This method can be overriden to customize the visitor behavior.

fn visit_jsx_expr_container(&mut self, n: &JSXExprContainer)

This method can be overriden to customize the visitor behavior.

fn visit_jsx_fragment(&mut self, n: &JSXFragment)

This method can be overriden to customize the visitor behavior.

fn visit_jsx_member_expr(&mut self, n: &JSXMemberExpr)

This method can be overriden to customize the visitor behavior.

fn visit_jsx_namespaced_name(&mut self, n: &JSXNamespacedName)

This method can be overriden to customize the visitor behavior.

fn visit_jsx_object(&mut self, n: &JSXObject)

This method can be overriden to customize the visitor behavior.

fn visit_jsx_opening_element(&mut self, n: &JSXOpeningElement)

This method can be overriden to customize the visitor behavior.

fn visit_jsx_opening_fragment(&mut self, n: &JSXOpeningFragment)

This method can be overriden to customize the visitor behavior.

fn visit_jsx_spread_child(&mut self, n: &JSXSpreadChild)

This method can be overriden to customize the visitor behavior.

fn visit_jsx_text(&mut self, n: &JSXText)

This method can be overriden to customize the visitor behavior.

fn visit_key_value_pat_prop(&mut self, n: &KeyValuePatProp)

This method can be overriden to customize the visitor behavior.

fn visit_key_value_prop(&mut self, n: &KeyValueProp)

This method can be overriden to customize the visitor behavior.

fn visit_labeled_stmt(&mut self, n: &LabeledStmt)

This method can be overriden to customize the visitor behavior.

fn visit_lit(&mut self, n: &Lit)

This method can be overriden to customize the visitor behavior.

fn visit_member_expr(&mut self, n: &MemberExpr)

This method can be overriden to customize the visitor behavior.

fn visit_member_prop(&mut self, n: &MemberProp)

This method can be overriden to customize the visitor behavior.

fn visit_meta_prop_expr(&mut self, n: &MetaPropExpr)

This method can be overriden to customize the visitor behavior.

fn visit_meta_prop_kind(&mut self, n: &MetaPropKind)

This method can be overriden to customize the visitor behavior.

fn visit_method_kind(&mut self, n: &MethodKind)

This method can be overriden to customize the visitor behavior.

fn visit_method_prop(&mut self, n: &MethodProp)

This method can be overriden to customize the visitor behavior.

fn visit_module(&mut self, n: &Module)

This method can be overriden to customize the visitor behavior.

fn visit_module_decl(&mut self, n: &ModuleDecl)

This method can be overriden to customize the visitor behavior.

fn visit_module_export_name(&mut self, n: &ModuleExportName)

This method can be overriden to customize the visitor behavior.

fn visit_module_item(&mut self, n: &ModuleItem)

This method can be overriden to customize the visitor behavior.

fn visit_module_items(&mut self, n: &[ModuleItem])

fn visit_named_export(&mut self, n: &NamedExport)

This method can be overriden to customize the visitor behavior.

fn visit_new_expr(&mut self, n: &NewExpr)

This method can be overriden to customize the visitor behavior.

fn visit_null(&mut self, n: &Null)

This method can be overriden to customize the visitor behavior.

fn visit_number(&mut self, n: &Number)

This method can be overriden to customize the visitor behavior.

fn visit_object_lit(&mut self, n: &ObjectLit)

This method can be overriden to customize the visitor behavior.

fn visit_object_pat(&mut self, n: &ObjectPat)

This method can be overriden to customize the visitor behavior.

fn visit_object_pat_prop(&mut self, n: &ObjectPatProp)

This method can be overriden to customize the visitor behavior.

fn visit_object_pat_props(&mut self, n: &[ObjectPatProp])

fn visit_opt_accessibility(&mut self, n: Option<&Accessibility>)

fn visit_opt_atom(&mut self, n: Option<&Atom>)

fn visit_opt_block_stmt(&mut self, n: Option<&BlockStmt>)

fn visit_opt_call(&mut self, n: &OptCall)

This method can be overriden to customize the visitor behavior.

fn visit_opt_catch_clause(&mut self, n: Option<&CatchClause>)

fn visit_opt_chain_base(&mut self, n: &OptChainBase)

This method can be overriden to customize the visitor behavior.

fn visit_opt_chain_expr(&mut self, n: &OptChainExpr)

This method can be overriden to customize the visitor behavior.

fn visit_opt_expr(&mut self, n: Option<&Box<Expr>>)

fn visit_opt_expr_or_spread(&mut self, n: Option<&ExprOrSpread>)

fn visit_opt_expr_or_spreads(&mut self, n: Option<&[ExprOrSpread]>)

fn visit_opt_ident(&mut self, n: Option<&Ident>)

fn visit_opt_jsx_attr_value(&mut self, n: Option<&JSXAttrValue>)

fn visit_opt_jsx_closing_element(&mut self, n: Option<&JSXClosingElement>)

fn visit_opt_module_export_name(&mut self, n: Option<&ModuleExportName>)

fn visit_opt_module_items(&mut self, n: Option<&[ModuleItem]>)

fn visit_opt_object_lit(&mut self, n: Option<&Box<ObjectLit>>)

fn visit_opt_pat(&mut self, n: Option<&Pat>)

fn visit_opt_span(&mut self, n: Option<&Span>)

fn visit_opt_stmt(&mut self, n: Option<&Box<Stmt>>)

fn visit_opt_str(&mut self, n: Option<&Box<Str>>)

fn visit_opt_true_plus_minus(&mut self, n: Option<&TruePlusMinus>)

fn visit_opt_ts_entity_name(&mut self, n: Option<&TsEntityName>)

fn visit_opt_ts_namespace_body(&mut self, n: Option<&TsNamespaceBody>)

fn visit_opt_ts_type(&mut self, n: Option<&Box<TsType>>)

fn visit_opt_ts_type_ann(&mut self, n: Option<&Box<TsTypeAnn>>)

fn visit_opt_ts_type_param_decl(&mut self, n: Option<&Box<TsTypeParamDecl>>)

fn visit_opt_ts_type_param_instantiation(
&mut self,
n: Option<&Box<TsTypeParamInstantiation>>
)

fn visit_opt_var_decl_or_expr(&mut self, n: Option<&VarDeclOrExpr>)

fn visit_opt_vec_expr_or_spreads(&mut self, n: &[Option<ExprOrSpread>])

fn visit_opt_vec_pats(&mut self, n: &[Option<Pat>])

fn visit_param(&mut self, n: &Param)

This method can be overriden to customize the visitor behavior.

fn visit_param_or_ts_param_prop(&mut self, n: &ParamOrTsParamProp)

This method can be overriden to customize the visitor behavior.

fn visit_param_or_ts_param_props(&mut self, n: &[ParamOrTsParamProp])

fn visit_params(&mut self, n: &[Param])

fn visit_paren_expr(&mut self, n: &ParenExpr)

This method can be overriden to customize the visitor behavior.

fn visit_pat(&mut self, n: &Pat)

This method can be overriden to customize the visitor behavior.

fn visit_pat_or_expr(&mut self, n: &PatOrExpr)

This method can be overriden to customize the visitor behavior.

fn visit_pats(&mut self, n: &[Pat])

fn visit_private_method(&mut self, n: &PrivateMethod)

This method can be overriden to customize the visitor behavior.

fn visit_private_name(&mut self, n: &PrivateName)

This method can be overriden to customize the visitor behavior.

fn visit_private_prop(&mut self, n: &PrivateProp)

This method can be overriden to customize the visitor behavior.

fn visit_program(&mut self, n: &Program)

This method can be overriden to customize the visitor behavior.

fn visit_prop(&mut self, n: &Prop)

This method can be overriden to customize the visitor behavior.

fn visit_prop_name(&mut self, n: &PropName)

This method can be overriden to customize the visitor behavior.

fn visit_prop_or_spread(&mut self, n: &PropOrSpread)

This method can be overriden to customize the visitor behavior.

fn visit_prop_or_spreads(&mut self, n: &[PropOrSpread])

fn visit_regex(&mut self, n: &Regex)

This method can be overriden to customize the visitor behavior.

fn visit_reserved_unused(&mut self, n: &ReservedUnused)

This method can be overriden to customize the visitor behavior.

fn visit_rest_pat(&mut self, n: &RestPat)

This method can be overriden to customize the visitor behavior.

fn visit_return_stmt(&mut self, n: &ReturnStmt)

This method can be overriden to customize the visitor behavior.

fn visit_script(&mut self, n: &Script)

This method can be overriden to customize the visitor behavior.

fn visit_seq_expr(&mut self, n: &SeqExpr)

This method can be overriden to customize the visitor behavior.

fn visit_setter_prop(&mut self, n: &SetterProp)

This method can be overriden to customize the visitor behavior.

fn visit_span(&mut self, n: &Span)

fn visit_spread_element(&mut self, n: &SpreadElement)

This method can be overriden to customize the visitor behavior.

fn visit_static_block(&mut self, n: &StaticBlock)

This method can be overriden to customize the visitor behavior.

fn visit_stmt(&mut self, n: &Stmt)

This method can be overriden to customize the visitor behavior.

fn visit_stmts(&mut self, n: &[Stmt])

fn visit_str(&mut self, n: &Str)

This method can be overriden to customize the visitor behavior.

fn visit_super(&mut self, n: &Super)

This method can be overriden to customize the visitor behavior.

fn visit_super_prop(&mut self, n: &SuperProp)

This method can be overriden to customize the visitor behavior.

fn visit_super_prop_expr(&mut self, n: &SuperPropExpr)

This method can be overriden to customize the visitor behavior.

fn visit_switch_case(&mut self, n: &SwitchCase)

This method can be overriden to customize the visitor behavior.

fn visit_switch_cases(&mut self, n: &[SwitchCase])

fn visit_switch_stmt(&mut self, n: &SwitchStmt)

This method can be overriden to customize the visitor behavior.

fn visit_tagged_tpl(&mut self, n: &TaggedTpl)

This method can be overriden to customize the visitor behavior.

fn visit_this_expr(&mut self, n: &ThisExpr)

This method can be overriden to customize the visitor behavior.

fn visit_throw_stmt(&mut self, n: &ThrowStmt)

This method can be overriden to customize the visitor behavior.

fn visit_tpl(&mut self, n: &Tpl)

This method can be overriden to customize the visitor behavior.

fn visit_tpl_element(&mut self, n: &TplElement)

This method can be overriden to customize the visitor behavior.

fn visit_tpl_elements(&mut self, n: &[TplElement])

fn visit_true_plus_minus(&mut self, n: &TruePlusMinus)

This method can be overriden to customize the visitor behavior.

fn visit_try_stmt(&mut self, n: &TryStmt)

This method can be overriden to customize the visitor behavior.

fn visit_ts_array_type(&mut self, n: &TsArrayType)

This method can be overriden to customize the visitor behavior.

fn visit_ts_as_expr(&mut self, n: &TsAsExpr)

This method can be overriden to customize the visitor behavior.

fn visit_ts_call_signature_decl(&mut self, n: &TsCallSignatureDecl)

This method can be overriden to customize the visitor behavior.

fn visit_ts_conditional_type(&mut self, n: &TsConditionalType)

This method can be overriden to customize the visitor behavior.

fn visit_ts_const_assertion(&mut self, n: &TsConstAssertion)

This method can be overriden to customize the visitor behavior.

fn visit_ts_construct_signature_decl(&mut self, n: &TsConstructSignatureDecl)

This method can be overriden to customize the visitor behavior.

fn visit_ts_constructor_type(&mut self, n: &TsConstructorType)

This method can be overriden to customize the visitor behavior.

fn visit_ts_entity_name(&mut self, n: &TsEntityName)

This method can be overriden to customize the visitor behavior.

fn visit_ts_enum_decl(&mut self, n: &TsEnumDecl)

This method can be overriden to customize the visitor behavior.

fn visit_ts_enum_member(&mut self, n: &TsEnumMember)

This method can be overriden to customize the visitor behavior.

fn visit_ts_enum_member_id(&mut self, n: &TsEnumMemberId)

This method can be overriden to customize the visitor behavior.

fn visit_ts_enum_members(&mut self, n: &[TsEnumMember])

fn visit_ts_export_assignment(&mut self, n: &TsExportAssignment)

This method can be overriden to customize the visitor behavior.

fn visit_ts_expr_with_type_args(&mut self, n: &TsExprWithTypeArgs)

This method can be overriden to customize the visitor behavior.

fn visit_ts_expr_with_type_args_vec(&mut self, n: &[TsExprWithTypeArgs])

fn visit_ts_external_module_ref(&mut self, n: &TsExternalModuleRef)

This method can be overriden to customize the visitor behavior.

fn visit_ts_fn_or_constructor_type(&mut self, n: &TsFnOrConstructorType)

This method can be overriden to customize the visitor behavior.

fn visit_ts_fn_param(&mut self, n: &TsFnParam)

This method can be overriden to customize the visitor behavior.

fn visit_ts_fn_params(&mut self, n: &[TsFnParam])

fn visit_ts_fn_type(&mut self, n: &TsFnType)

This method can be overriden to customize the visitor behavior.

fn visit_ts_getter_signature(&mut self, n: &TsGetterSignature)

This method can be overriden to customize the visitor behavior.

fn visit_ts_import_equals_decl(&mut self, n: &TsImportEqualsDecl)

This method can be overriden to customize the visitor behavior.

fn visit_ts_import_type(&mut self, n: &TsImportType)

This method can be overriden to customize the visitor behavior.

fn visit_ts_index_signature(&mut self, n: &TsIndexSignature)

This method can be overriden to customize the visitor behavior.

fn visit_ts_indexed_access_type(&mut self, n: &TsIndexedAccessType)

This method can be overriden to customize the visitor behavior.

fn visit_ts_infer_type(&mut self, n: &TsInferType)

This method can be overriden to customize the visitor behavior.

fn visit_ts_instantiation(&mut self, n: &TsInstantiation)

This method can be overriden to customize the visitor behavior.

fn visit_ts_interface_body(&mut self, n: &TsInterfaceBody)

This method can be overriden to customize the visitor behavior.

fn visit_ts_interface_decl(&mut self, n: &TsInterfaceDecl)

This method can be overriden to customize the visitor behavior.

fn visit_ts_intersection_type(&mut self, n: &TsIntersectionType)

This method can be overriden to customize the visitor behavior.

fn visit_ts_keyword_type(&mut self, n: &TsKeywordType)

This method can be overriden to customize the visitor behavior.

fn visit_ts_keyword_type_kind(&mut self, n: &TsKeywordTypeKind)

This method can be overriden to customize the visitor behavior.

fn visit_ts_lit(&mut self, n: &TsLit)

This method can be overriden to customize the visitor behavior.

fn visit_ts_lit_type(&mut self, n: &TsLitType)

This method can be overriden to customize the visitor behavior.

fn visit_ts_mapped_type(&mut self, n: &TsMappedType)

This method can be overriden to customize the visitor behavior.

fn visit_ts_method_signature(&mut self, n: &TsMethodSignature)

This method can be overriden to customize the visitor behavior.

fn visit_ts_module_block(&mut self, n: &TsModuleBlock)

This method can be overriden to customize the visitor behavior.

fn visit_ts_module_decl(&mut self, n: &TsModuleDecl)

This method can be overriden to customize the visitor behavior.

fn visit_ts_module_name(&mut self, n: &TsModuleName)

This method can be overriden to customize the visitor behavior.

fn visit_ts_module_ref(&mut self, n: &TsModuleRef)

This method can be overriden to customize the visitor behavior.

fn visit_ts_namespace_body(&mut self, n: &TsNamespaceBody)

This method can be overriden to customize the visitor behavior.

fn visit_ts_namespace_decl(&mut self, n: &TsNamespaceDecl)

This method can be overriden to customize the visitor behavior.

fn visit_ts_namespace_export_decl(&mut self, n: &TsNamespaceExportDecl)

This method can be overriden to customize the visitor behavior.

fn visit_ts_non_null_expr(&mut self, n: &TsNonNullExpr)

This method can be overriden to customize the visitor behavior.

fn visit_ts_optional_type(&mut self, n: &TsOptionalType)

This method can be overriden to customize the visitor behavior.

fn visit_ts_param_prop(&mut self, n: &TsParamProp)

This method can be overriden to customize the visitor behavior.

fn visit_ts_param_prop_param(&mut self, n: &TsParamPropParam)

This method can be overriden to customize the visitor behavior.

fn visit_ts_parenthesized_type(&mut self, n: &TsParenthesizedType)

This method can be overriden to customize the visitor behavior.

fn visit_ts_property_signature(&mut self, n: &TsPropertySignature)

This method can be overriden to customize the visitor behavior.

fn visit_ts_qualified_name(&mut self, n: &TsQualifiedName)

This method can be overriden to customize the visitor behavior.

fn visit_ts_rest_type(&mut self, n: &TsRestType)

This method can be overriden to customize the visitor behavior.

fn visit_ts_satisfies_expr(&mut self, n: &TsSatisfiesExpr)

This method can be overriden to customize the visitor behavior.

fn visit_ts_setter_signature(&mut self, n: &TsSetterSignature)

This method can be overriden to customize the visitor behavior.

fn visit_ts_this_type(&mut self, n: &TsThisType)

This method can be overriden to customize the visitor behavior.

fn visit_ts_this_type_or_ident(&mut self, n: &TsThisTypeOrIdent)

This method can be overriden to customize the visitor behavior.

fn visit_ts_tpl_lit_type(&mut self, n: &TsTplLitType)

This method can be overriden to customize the visitor behavior.

fn visit_ts_tuple_element(&mut self, n: &TsTupleElement)

This method can be overriden to customize the visitor behavior.

fn visit_ts_tuple_elements(&mut self, n: &[TsTupleElement])

fn visit_ts_tuple_type(&mut self, n: &TsTupleType)

This method can be overriden to customize the visitor behavior.

fn visit_ts_type(&mut self, n: &TsType)

This method can be overriden to customize the visitor behavior.

fn visit_ts_type_alias_decl(&mut self, n: &TsTypeAliasDecl)

This method can be overriden to customize the visitor behavior.

fn visit_ts_type_ann(&mut self, n: &TsTypeAnn)

This method can be overriden to customize the visitor behavior.

fn visit_ts_type_assertion(&mut self, n: &TsTypeAssertion)

This method can be overriden to customize the visitor behavior.

fn visit_ts_type_element(&mut self, n: &TsTypeElement)

This method can be overriden to customize the visitor behavior.

fn visit_ts_type_elements(&mut self, n: &[TsTypeElement])

fn visit_ts_type_lit(&mut self, n: &TsTypeLit)

This method can be overriden to customize the visitor behavior.

fn visit_ts_type_operator(&mut self, n: &TsTypeOperator)

This method can be overriden to customize the visitor behavior.

fn visit_ts_type_operator_op(&mut self, n: &TsTypeOperatorOp)

This method can be overriden to customize the visitor behavior.

fn visit_ts_type_param(&mut self, n: &TsTypeParam)

This method can be overriden to customize the visitor behavior.

fn visit_ts_type_param_decl(&mut self, n: &TsTypeParamDecl)

This method can be overriden to customize the visitor behavior.

fn visit_ts_type_param_instantiation(&mut self, n: &TsTypeParamInstantiation)

This method can be overriden to customize the visitor behavior.

fn visit_ts_type_params(&mut self, n: &[TsTypeParam])

fn visit_ts_type_predicate(&mut self, n: &TsTypePredicate)

This method can be overriden to customize the visitor behavior.

fn visit_ts_type_query(&mut self, n: &TsTypeQuery)

This method can be overriden to customize the visitor behavior.

fn visit_ts_type_query_expr(&mut self, n: &TsTypeQueryExpr)

This method can be overriden to customize the visitor behavior.

fn visit_ts_type_ref(&mut self, n: &TsTypeRef)

This method can be overriden to customize the visitor behavior.

fn visit_ts_types(&mut self, n: &[Box<TsType>])

fn visit_ts_union_or_intersection_type(&mut self, n: &TsUnionOrIntersectionType)

This method can be overriden to customize the visitor behavior.

fn visit_ts_union_type(&mut self, n: &TsUnionType)

This method can be overriden to customize the visitor behavior.

fn visit_unary_expr(&mut self, n: &UnaryExpr)

This method can be overriden to customize the visitor behavior.

fn visit_unary_op(&mut self, n: &UnaryOp)

This method can be overriden to customize the visitor behavior.

fn visit_update_expr(&mut self, n: &UpdateExpr)

This method can be overriden to customize the visitor behavior.

fn visit_update_op(&mut self, n: &UpdateOp)

This method can be overriden to customize the visitor behavior.

fn visit_var_decl(&mut self, n: &VarDecl)

This method can be overriden to customize the visitor behavior.

fn visit_var_decl_kind(&mut self, n: &VarDeclKind)

This method can be overriden to customize the visitor behavior.

fn visit_var_decl_or_expr(&mut self, n: &VarDeclOrExpr)

This method can be overriden to customize the visitor behavior.

fn visit_var_decl_or_pat(&mut self, n: &VarDeclOrPat)

This method can be overriden to customize the visitor behavior.

fn visit_var_declarator(&mut self, n: &VarDeclarator)

This method can be overriden to customize the visitor behavior.

fn visit_var_declarators(&mut self, n: &[VarDeclarator])

fn visit_while_stmt(&mut self, n: &WhileStmt)

This method can be overriden to customize the visitor behavior.

fn visit_with_stmt(&mut self, n: &WithStmt)

This method can be overriden to customize the visitor behavior.

fn visit_yield_expr(&mut self, n: &YieldExpr)

This method can be overriden to customize the visitor behavior.