Crate swc_ecma_visit

Macros

assert_eq_ignore_span

noop_fold_type

Note: Ignoring more types is not considered as a breaking change.

noop_visit_mut_type

Note: Ignoring more types is not considered as a breaking change.

noop_visit_type

Note: Ignoring more types is not considered as a breaking change.

standard_only_fold

Mark visitor as ECMAScript standard only and mark other types as unreachable.

standard_only_visit

Mark visitor as ECMAScript standard only and mark other types as unreachable.

standard_only_visit_mut

Mark visitor as ECMAScript standard only and mark other types as unreachable.

visit_mut_obj_and_computed

visit_obj_and_computed

Structs

All

Visit all children nodes. This converts VisitAll to Visit. The type parameter V should implement VisitAll and All<V> implements Visit.

FoldPass

VisitMutPass

VisitPass

Enums

NodeRef

Traits

Fold

A visitor trait for traversing the AST.

FoldWith

A trait implemented for types that can be visited using a visitor.

InjectVars

Implemented for passes which inject variables.

Visit

A visitor trait for traversing the AST.

VisitMut

A visitor trait for traversing the AST.

VisitMutWith

A trait implemented for types that can be visited using a visitor.

VisitWith

A trait implemented for types that can be visited using a visitor.

Functions

fold_pass

span_remover

Returns a Fold which changes all span into DUMMY_SP.

visit_mut_pass

visit_pass