Trait swc_ecma_visit::InjectVars

pub trait InjectVars {
    fn take_vars(&mut self) -> Vec<VarDeclarator>;
}

Implemeented for passes which inject varaibles.

If a pass depends on other pass which injects variables, this trait can be used to keep the varaibles.

Required Methods

fn take_vars(&mut self) -> Vec<VarDeclarator>